backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.0.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.33.3
filelock==3.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
fsspec==2025.5.1
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.5
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.0.1
wsproto==1.2.0
